# chore: integration tests performance fixes