--- a/backend/data/build.gradle
+++ b/backend/data/build.gradle
@@ -218,6 +218,8 @@ sourceSets {
 }
 
 dependencyManagement {
+    applyMavenExclusions = false
+
     imports {
         mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
     }