--- a/backend/development/build.gradle
+++ b/backend/development/build.gradle
@@ -80,6 +80,8 @@ sourceSets {
 }
 
 dependencyManagement {
+    applyMavenExclusions = false
+
     imports {
         mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
     }