--- a/backend/security/build.gradle
+++ b/backend/security/build.gradle
@@ -84,6 +84,7 @@ sourceSets {
 }
 
 dependencyManagement {
+    applyMavenExclusions = false
     imports {
         mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
     }