--- a/backend/testing/build.gradle
+++ b/backend/testing/build.gradle
@@ -142,6 +142,8 @@ sourceSets {
 }
 
 dependencyManagement {
+    applyMavenExclusions = false
+
     imports {
         mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
     }