--- a/ee/backend/app/build.gradle
+++ b/ee/backend/app/build.gradle
@@ -96,6 +96,8 @@ kotlin {
 }
 
 dependencyManagement {
+    applyMavenExclusions = false
+
     imports {
         mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
     }