--- a/ee/backend/tests/build.gradle
+++ b/ee/backend/tests/build.gradle
@@ -61,6 +61,8 @@ kotlin {
 }
 
 dependencyManagement {
+    applyMavenExclusions = false
+
     imports {
         mavenBom org.springframework.boot.gradle.plugin.SpringBootPlugin.BOM_COORDINATES
     }