--- a/gradle.properties
+++ b/gradle.properties
@@ -10,3 +10,4 @@ org.gradle.parallel=true
 jacksonVersion=2.13.5
 slackSdkVersion=1.37.0
 commonsLang3Version=3.14.0
+org.gradle.caching=true