@JanCizmar 2024-08-20
> Gradle configuration phase takes around 3 sec, from which 90% takes Sentry.

- [Can you maybe look into this? Update it or create issue in their github?

dependencyResolution phase took 4-7 seconds. I've fixed it around 1 sec with https://github.com/spring-gradle-plugins/dependency-management-plugin/issues/222#issuecomment-411005109

- Thanks!

Task kaptTestKotlin takes at least 1.5 seconds and misses cache every time. Consider upgrade to KSP and to Kotlin 2.0. I excluded build-info.properties that causes cache miss through Gradle Cache normalization, but exclusion doesn't seem to work consistently. Is Spring.buildInfo() call important in this project or is just a leftover?

> O don't remember adding `Spring.buildInfo() `, you can remove it. If it helps. You can also try to update to kotlin 2.0. If tests pass, we're fine.

> backend/data/build/resources/main/sentry-debug-meta.properties causes cache miss every time as well, adding 0.5 sec. I couldn’t exclude this file from Gradle Cache hit computation using the same normalisation config.

- So what now?

Sometimes build (without clean) takes 1min. Not sure if this issue happens during normal development as well, but I reckon it does, making negative impact on developers experience.

- Let's see...

The log output is heavily polluted (e.g. by Hibernate queries dump). It affects execution performance, though currently the impact shouldn't be very high.

- Yeah, maybe we can prevent hibernate from logging for tests. However, it's useful when we write tests so we can spot n+1 or too long queries.